+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.2.3 (2026-01-10)
+
+- [NEW] Configure output directory.
+- [NEW] Check M3U file validity.
+
 ## v0.2.2 (2026-01-09)
 
 - [NEW] If not a `master.m3u8` found, allow to download the content anyway.